requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.0
